requirements.txt
python-dotenv
langchain
ipykernel
# unstructured[all-docs]
huggingface-hub
sentence_transformers
streamlit
pypdf
langchain-google-genai